Changes:
Comments to separate different sections of the code are more visible. There's still alot of commenting that can be improved though.
Guest now implements Serializable. (not sure why it wasn't before)
logic.setEventStartDate and setEventEndDate does not add 1 and 1900 to month and year now. When creating a Date object, setting year is eg. 2012 will cause the dateChooser to reflect 2012 + 1900.
logic.setDateList now generates the list correctly. Previous implementation did not work for different months.
Updated getProgrammeSchedule() to include the new Date field introduced.
startTime and endTime is changed back to -1 by default. This is because Midnight is 0000.
If startTime and endTime is -1, the GUI does not show any value now.