Fixed bug where meal date selected isn't saved.
Problem: Whenever a saved file is loaded, updateStep1 is called. This will set the dates, which will in turn fire property changes in the dates, which will generate the combo box dates. However, this generation will fire action events in the combo box, causing it to set the index back to -1 (when all items are removed), and 0 when items are placed back.
Solution: Remove the action listener of comboBox during the generation of combo box dates, and add it back after the generation is done. This is to avoid overwriting the already selected index.

I have sort of tested this, but do play with it more. This firing of events are tricky.